Rename ann_ivf.proto to ann.proto and knn_proto.rs to ann_proto.rs
